Fixing free when swapping
On GVT, we might swap malloc areas to keep the metadata compacted. In that case, the per-chunk pointer used for free's might be uncorrectly set. This is solved by having each malloc area keep a self pointer. The per-chunk pointer points to the self pointer. If we swap, we simply update it.