configuration summary
When running `./configure`, at the end we display a summary of the
current configuration on the machine.
This allows to keep track of what subsystems will be active, and
shows the user a direct choice to disable/enable some of them,
telling what are the required dependencies if not satisfied.