Statistics on GVT are buffered in memory
This allows us to prevent journald to take control on journaled
file systems. Preliminary experiments tell that the CPU requirement
of journald can be as high as 30% in some cases. This should reduce
the number of rollbacks in general, and avoid receiving a high number
of interrupts when using libtimestretch, which therefore would force
the platform to miss a non-minimal number of extra ticks.