Fixing GVT and Termination critical race
A critical race can be encountered when specifying a certain amount
of simulation time after which the simulation has to be halted.
In particular, the early adoption of the new GVT value in the aware
phase of the non-blocking GVT algorithm might prevent some other thread
from reaching the end phase, thus the termination will never occur and
the simulation will hang.
This is solved by adding an additional per-thread variable which is
used to keep the freshly-adopted GVT value, but without making it
accessible to the get_last_gvt() function.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>