Handling arrays and structs - preliminary
Global variables larger than 8 bytes are now handles as a set of
n different 8-bytes variables.
Note that in this way, the write/read procedures might fail, for
example in case one 8-byte subset of the larger global variable
is actually composed of multiple smaller-sized variables.
In this case the write/read procedures might not be able to
identify the correct list where to apply the operation.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>