Fix bug in generation of allocated chunk in buddy
The used buddy systems works on a chunk granularity basis. When making
requests to the buddy system, the allocator wrapper was correctly using
a conversion from bytes to chunks. This was not done when returning the
actual pointer. In this way, subsequent allocations where given
overlapped regions of memory, which resulted in inconsistent memory
management, which in its turn was squashing malloc() when allocating
buffers for log states, which where overflowing the buffers.

Additionally, this commit changes some internal configuration parameters
of DyMeLoR, so that internal fragmentation to the buddy system is kept
very low.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>