Working version of overtick
The main problem was that if an interrupt was received from a leaf
function, the kernel module could have written the return value in
the system V x86_64 ABI's red zone. In this case, the return value
could have been incorrectly used by leaf functions as memory addresses
where variables local to the functions where stored.
To fix this, -mno-red-zone becomes a fundamental compilation flag
for ROOT-Sim.

This commit includes as well some minor optimizations and small
inconsistencies' fixes in the code.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>