Chunk dumping area was not freed
The dumping area for the chunk reversal, was not freed once malloc'd.
A simple handling of this case has been added by adding a new field in
the revwin structure in order to keep track of the dump address.

Functions affected are execute_undo_event and revwin_free, which handle
this case.