Fixed while condition in reverse rollback cycle
The first while cycle in the reverse branch of the rollback
procedure check whether the object is NULL only after an access
tempative of one field of the same object, hence resulting in
a possible segmentation fault.

It has been simply swapped the two AND operands of the inner condition.

Signed-off-by: Davide Cingolani <cingodvd@gmail.com>