Wrap abort() to dump information on the event
stdlib wrapper now wraps abort() as well. If abort() was called by the
application-level code, or by the platform during the execution of an
event, if ROOT-Sim is compiled with --enable-debug a dump of the content
of the event is printed on screen.
This is based on a cherry-picked commit from timestretch branch.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>
(cherry picked from commit 9c93918fa00b70c8d12d966fb27eab1d26851d6c)