bugfix: rootsim-cc only works in debug mode
Bug
---
Some flags were missing in the CFLAGS variable when the
debug mode was not active.
Since some of these flags are required, the compiled `rootsim-cc`
wasn't working when the debug mode is disabled.

Fix
---
- The conditional variable DEBUG is now setup and can be used also in
the Makefile
- All the required compilation flags are stored in `CFLAGS_COMMON` and
are used both when debug mode is active as well when is disabled.

Signed-off-by: ael-code <tommy.ael@gmail.com>