bugfix: rootsim-cc does not clean temp files
Bug
---
In the cleanup phase it was wrongly assumed that the model sources
reside in the current working direcotry. Moreover temporary
files not produced in current session were deleted.

Fix
---
The temporary files are correctly detected using the `SOURCES` file array.
Only the temporary files produced in the current session will be deleted