current_lp is not a gid
current_lp is a thread variable and should always points to the current LP being scheduled by the current thread. In the case no LP is being scheduled it should point to IDLE_PROCESS.

Reasobably it should be the global id of the LP (gid), but at the moment for practical and historical reasons it points to the local id of the current LP (lid).

The problem is that in some point of the code is treated as a gid even if it is a lid.

This commit fix all the code points in which current_lp is treated as local LP id