new cli flag: --no-core-binding
Added new cli flag that can be used to disable the binding of the
threads to the phisical processing cores. It could be useful in the case
other external binding approach need to be used and in the case
that more then one kernel is executing on the same machine.