ECS reworking
This is a partial commit, where ECS has been mostly realigned to the
current state of the simulation platform.
There are many printf's for debugging purposes, which will be later
removed.

There is still a corner case when the degree of parallelism is extremely
high, such that a log related to ECS is not taken, before a rollback
operation takes places.

Signed-off-by: Matteo Principe <matteo.principe92@gmail.com>