Implementation of generic lists
The basic implementation of a list relies on per-LP memory allocation to
keep all buffers separated across different LPs.

There has been the need, for the MPI subsystem, to keep generic lists
which are handled on a non-per-LP basis.

This commit introduces a variant of the list which allows to allocate
generic non-per-LP lists.
The cost of this implementation is that there are additional if's on the
critical path in all primitives which allocate memory. Although the
checks are on integer variables, and should be therefore quite quick, I
did not make any actual performance analysis of the impact on the
throughput of the simulation. I feel anyhow confident that this does not
affect the overall efficiency of the simulation engine significantly.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>