Rewrite MPI configuration rules
In this commit, we don't simply check for the availability of an
MPI compiling toolchain: we also check if the installed library
is recent enough to support asynchronous primitives which are
used to implement the GVT reduction algorithm.

This fixes #60.

Moreover, the `configure` switch `--with-mpi` has been renamed
`--enable-mpi` to make it consistent with the other switches.

Some minor polishing of `configure.ac` is as well introduced here.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>