Fix MPI check
The previous MPI check in `configure.ac` was buggy in a twofold way.
First, it was assuming that `cc` or `gcc` are valid MPI compilers,
which is certainly not true.

Then, it was emitting a message regarding a test of asynchronous
primitives even in the case MPI support was not enabled.

Finally, I've updated the configuration message so that if MPI support
is enabled but no suitable MPI compiler is found, it tells so.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>