Complete zero-copy message passing
The list implementation has been modified so as to avoid any copy of its
paylod. This required a significant reengineering of the module, which
now implements generics only relying on macro.

Memory deallocation of nodes has been solved via "function pointers" in
the only survived member which pruned a portion of the list.

The code base has been realigned to this change.

This fixes #65.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>