Cannot validate `idle_process`
As per issue #74, `idle_process` does not have a legal id---it is
actually set to `UINT_MAX`.

This was causing trouble when a model doesn't have a symmetric load, and
therefore a thread was trying to "schedule" the idle process. This
trigger caused the statistics subsystem to be activated which, in turn,
was checking for a legal lid value.

This is intended as a quick patch, as issue #66 should make avoid
relying on the `LID_t` type on a large part of the code base.

Fixes #74.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>