Fix a race condition on TCBs
The incarnation of an asymmetric thread is deterministically stored into
a Thread Control Block.

Upon the creation of a thread, a non-blocking agreement across the
threads is run, to let each thread pick a TCB independently of the
others. Here we were allocating the `tid` of the thread.

The tid of the threads was anyhow used statically to associate PTs to
CTs. The scenario was that a CT might change nature (and become a PT)
due to the non-blocking allocation. Yet, all the other data structures
did not change.

Here we associate the `tid` to the TCB deterministically. When a thread
picks a TCB via the non-blocking agreement, this determines both its
incarnation and its `tid` value.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>