Power: added new batch scheduling algorithm
Introduced a batch scheduling algorithm that should
significantly reduce the time necessary to schedule
a batch of events to PTs.

It inserts a parametric number of events for each LP
(with no full ports) to a priority queue with timestamp
as priority. Afterwards, it dequeues events in order,
updates port_events_to_fill and sends the element
in the input port of the respective PT

It is not the default mode, which is still stf.
It can be turned on by setting the execution time
parameter --scheduler batch.

Signed-off-by: Stefano Conoci <conoci@diag.uniroma1.it>