Fix astar compare function
The function to compare the astar_t type was returning an unproper
order of the elements, since the UINT_MAX cast to int which flips
the sign.

This commit introduces other minor changes introducing some sanity
checks and a new API function to check whether the a given cell is
marked as obstacle or not.

Signed-off-by: Davide Cingolani <cingolani@diag.uniroma1.it>