Add new event type and API functions
This commit adds a new event type AGENT_CHANGE_DEST to be indended to possibly
change the destination of an agent before it has reached the current destination.

A new API function `agent_new_destination()` is introduced in order to compute
the path to the next destination and fill the agent's descriptor properly.

Other minor changes has been introduced in the application main file:
* Agents are handled always by copying the content of the event, without the double
  malloc/memcpy passage via a temporary agent descriptor;
* Comments to the agent_t type has been added;

Signed-off-by: Davide Cingolani <cingolani@diag.uniroma1.it>