Fixed MPI multithreading issue
When running in an MPI multithreading environment, MPI_Iprobe + MPI_Recv
do not guarantee thread safety. Thus, when receiving messages with different
sizes (as we need for page prefetching), we could allocate a wrong amount
of memory for the incoming message. This was a knonw issue in MPI-2,
which was fixed in MPI-3 with the calls used in this commit. To fix this
in MPI-2, we should use a mutex. It is still a TODO.