Silence ar warning during make
Silence warning: ar: 'u' modifier ignored since 'D' is the default
This has to do with libtool. Different versions of libtool rely on
different macros, so we set the flags to the correct `cr` value
in both AR_FLAGS and ARFLAGS. AR_FLAGS is currently deprecated,
so it should go away at some point.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>