Significantly simplify the initialization of ULTs
We decided to give up with portability and have an x86-only implementation.
Then, the creation of a new ULT can be incredibly simpler, provided that
a ULT entry point will never-ever return.

The new organization of ULTs is lighter (fixes #103) and does not break
any calling convention (fixes #91). Therefore they can also be used when
using compiler optimization (fixes #92).

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>