Fix a bug in atomic.h and realign non-blocking algorithms
This commit fixes a small bug in atomic.h in which the old semantic
of cmpxcgh was used. This was creating a bug in the function
`spin_trylock`, such that it appeared as the lock was never taken.
This prevented the distributed GVT implementation to run, and any
thread to consume messages from the underlying MPI library.

Additionally, non-blocking algorithms have been modified so as to use
the relaxed memory model, rather than sequential consistency.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>