Simple SIGINT handling, improved MPI prints
Sending SIGINT to a ROOT-Sim process now causes its graceful termination.
Fixed a regression in which slave processes printed the configuration and the termination message.

There's still the issue of the models printing on all the machines their configuration (if they do so from ARGP_SUCCESS).
I'm not sure either that termination of a kernel is broadcasted to all MPI kernels in a distributed setting.

Criticism and suggestions are always welcome!

Signed-off-by: Andrea Piccione <piccions@gmx.com>