Update continuous integration script
This commit refactors the configuration of the Travis CI build.
In particular, most of the commands have been moved outside of
the .travis.yml file, into dedicated shell scripts in ./scripts.

The generation of the documentation now takes into account the fact
that we take some pages from the wiki and republish them in the
automatic commit in the website.

A new tests script has been written, which dumps a fancy table
with the results of each single test.

Due to update in .travis.yml, there is a mismatch between the
version of gcc used for compiling (gcc 8) and the version of gcov
used for code coverage (5 by default). A quick fix tells codecov
bash script to use gcov-8.

.travis.yml is now configured to use the most up-to-date
Ubuntu distro for testing, as older version of MPICH have bugs
which prevent to run correctly.

Slight changes to fix some doxygen warnings also have been included
in this commit.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>