Pass pointers to states rather than LIDs
This commit revolutionizes the simulation core. LID_t is no longer
the preferred way to pass information about the LPs which are being
handled by functions.

Rather, a struct lp_struct * is passed everywhere (it was LP_State *
before). This allows to remove a lot of redundancy, in which several
places of the simulator were keeping a pointer to the state, from
which the value of the LID was taken and passed as a parameter, only
to later retrieve again the pointer to the lp state.
This significantly increases code readability, simplifies development
and increases performance.

Two new macros have been introduced: foreach_lp and foreach_bound_lp.
They allow to implement loops over the LPs, automatically assigning
a correct pointer to the specified variable, which lives only in the
scope of the loop.
Their implementation is a bit sketchy: they use a per-thread variable
to iterate over the global array keeping pointers. This introduces a
memory access error (off by one in the array), which never causes a bug
as this is the actual ending condition of the loop. Anyhow, this has
significantly dirtied the output by valgrind. This will be fixed when
addressing issue #121, removing the need to iterate over a fixed array.

Most of the datastructures related to LPs have been moved into the
lp_struct structure, with two notable exceptions:

* memory map implemented by DyMeLoR. This requires non-minimal effort
  which makes sense to be done while refactoring the memory manager.
  This work is being carried out in a different branch, so it wasn't
  sane to mess that work up
* statistics: this is a long and tedious work, which could be done
  later. Anyhow, that module will significantly benefit in readability
  once this work is completed.

All in all, this huge code rewrite has fixed #66.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>