Introduce Unit Test Support
This commit fixes some warning in the dymelor Unit Test.
Similarly, the "packet" benchmark was hardcoded to run
using ECS (which is currently disabled), thus it was crashing
when using MPI.

Additionally, the testing script has been modified in order
to also run a batch of unit tests (currently composed only of
one test).

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>