Parallel allocator is the only allocator
This commit removes the HAVE_PARALLEL_ALLOCATOR macro and related stuff
from the code base.

In particular, given the new organization of the mm subsystem, it is pointless
to allow disabling it, as all features of the simulator currently depend
on it.

This also simplifies a bit the organization of the code.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>