Fixing Codecov and Travis Integration
Apparently, the gcov version shipped with g++-8 package on Travis
(gcov 8.0.1) has a bug which prevents it to correctly generate
coverage reports. This has been fixed in later versions, but it
is impossible to upgrade the package on Travis.

With this commit, we downgrade the compiling toolchain to gcc 7, which
brings back code coverage and supports the features we need to
test everything.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>