Fixed the foreach LP macros, ABM checkpointing.
The valgrind complaints about the foreach LP
macros have been fixed.

ABM now doesn't rely on the model's recoverable
memory: this required various changes in the hash_map
and in the agents instantiation and deletion.

Sadly enough, a disrupting bug which I think is
related with the rollback of the INIT message or
some other 0 timed message, provokes the crash
of the bug_explore model, since it relies heavily
on neighbours updates.

I will hopefully investigate the issue very quickly.

Criticism and suggestions are always welcome!

Signed-off-by: Andrea Piccione <piccions@gmx.com>