Align sequential termination detection
Termination detection in sequential simulation now behaves consistently
with the different termination detection schemes which are implemented
in the parallel runtime environment.

We have added the `accurate` termination detection which is not yet
implemented in the parallel simulator, as this requires a bit more of
engineering and will be added later on.

This fixes #122

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>