Adapted IME functionalities to RootSim module.
This commit provides an lightweight version of IME for RootSim.
It is present a set of utility methods for the configuration of
the PMCs (with PEBS support) as well as the customization of the
system IDT. Provisionally, the RootSim module is configured to
install a new IDT with two handlers (toggle_PMU and handle_PMI)
whose IRQ vector is hardcoded.
These must be defined in a programmatic way in the future.
IME core is shipped with a per_cpu buffer pool which is used to
copy gathered PEBS data when the buffer becomes full. This version
of the module does not take advantage of such a poll, but uses
a single big buffer. This can be read via IOCTl command, but it
is not thread-safe, hence it must be accessed after the profiling has
been stopped (done). Buffer is refreshed when it has been fully read.

Signed-off-by: Stefano Carna <stefano.carna.dev@gmail.com>