Backporting new LLVM-based compiler
This commit backports the new LLVM-based compiler into the main ROOT-Sim
repository.

Incidentally, to this end, we are now compiling the whole project with
LLVM rather than gcc.
Due to this, and given that we are using a large number of
GNU-extensions in the code base, this might be severely unstable.

To this end, also the new large number of warnings appearing in the code
base should be double checked. They show up only when configuring with
`--enable-debug`.

So long gcc, and thanks for all the headaches!

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>