Simplified version of PCS
This commit cleans up the current implementation of PCS, and gets rid of
unneeded linked data structures.
The codebase is cleaner and neater, and it's also performing faster,
which is a way to stress much more the underlying parallel runtime.

This is based on a partial cherry pick from:
b2a812b879fd85cce7892a73d2a8ce30f912f9f9
8bdcf2e77d27be99ecce18ce90cb00e57c74379f

Signed-off-by: Matteo Principe <matteo.principe92@gmail.com>
Signed-off-by: Andrea Piccione <piccions@gmx.com>
Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>