Add --silent-output runtime option
When running a model specifying `--silent-output`, any output generated
by both the model and the runtime is suppressed.

This is useful in production, when one is interested only in simulation
results (e.g. performance data in the output folder) or in simulation
results related to the committed portion, which should be saved to file
in `OnGVT()`.

This closes #2.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>