Fix clang warnings
Switching to clang emerged a large number of compiler warnings. These
were mostly don't cares, as the things were correct, but I loved how
clang is pedantic and highlights some points which could generate very
bad bad bad runtime errors.

One warning was related to the extraction of the value of the stack
pointer by relying on a GNU-C Register Extension which we used to read
the value of the stack pointer. Given that we don't really need
*exactly* the stack pointer, but the current function frame is ok as
well, this has been changed to the more portable compiler intrinsic
__builtin_frame_address(0).

Incidentally, some code quality improvements according to the project
standards have been introduced by this commit.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>