prepare for autonomic rollback mode
This commit introduces the code needed to support an autonomic switch between approximated and precise rollback mode. We
decided to implement this as an enum in order to let the modeler force a specific mode (namely, APPROXIMATED, PRECISE and
AUTONOMIC) wherever she wants in the model code. This could not be the best choice so I'm open to suggestions!
The autonomic switch is not implemented yet, and for now a dummy function was provided.

Signed-off-by: readbeard <matteo.principe92@gmail.com>