Enhancing statistics
This commit introduces new statistics data useful to compute the
formula for the autonomic approximated rollback scheme. Unfortunately,
due to the 'macro' nature of lists, I had to duplicate code for counting
committed approximated rollbacks. Note that the switches in the
statistics.c file where enhanced too, even if some of them aren't really
used: this was done for future compatibility and code completeness.

Signed-off-by: readbeard <matteo.principe92@gmail.com>