Introduce new (approximated) version of TBC
This commit introduces a new version of TBC in which the core portion of
the state is augmented with the number of agents in all the different
states (recall that in the standard approximated implementation we only
kept track of the counter of the healthy agents). Thus, this
implementation includes also a new version of RestoreApproximated taking
into account the LP state expansion.

Signed-off-by: readbeard <matteo.principe92@gmail.com>